// src/components/social-button/types.ts
export type SocialProvider = 
    | 'google' | 'facebook' | 'github' | 'twitter' | 'linkedin' | 'apple' 
    | 'microsoft' | 'discord' | 'slack' | 'spotify' | 'amazon' | 'pinterest' 
    | 'tiktok' | 'paypal' | 'twitch' | 'instagram' | 'youtube' | 'whatsapp' 
    | 'snapchat' | 'behance' | 'dribbble'